+++ src/slave/slave.cpp
@@ -1265,7 +1265,8 @@ void Slave::registered(
   CHECK_SOME(master);
 
   if (connection.has_total_ping_timeout_seconds()) {
-    masterPingTimeout = Seconds(connection.total_ping_timeout_seconds());
+    masterPingTimeout =
+      Seconds(static_cast<int64_t>(connection.total_ping_timeout_seconds()));
   } else {
     masterPingTimeout = DEFAULT_MASTER_PING_TIMEOUT();
   }
@@ -1368,7 +1369,8 @@ void Slave::reregistered(
   }
 
   if (connection.has_total_ping_timeout_seconds()) {
-    masterPingTimeout = Seconds(connection.total_ping_timeout_seconds());
+    masterPingTimeout =
+      Seconds(static_cast<int64_t>(connection.total_ping_timeout_seconds()));
   } else {
     masterPingTimeout = DEFAULT_MASTER_PING_TIMEOUT();
   }
@@ -7787,7 +7789,7 @@ Future<ResourceUsage> Slave::usage()
         // 'future' and 'executor' below should be in sync.
         CHECK_EQ(futures.size(), (size_t) usage->executors_size());
 
-        size_t i = 0;
+        int i = 0;
         foreach (const Future<ResourceStatistics>& future, futures) {
           ResourceUsage::Executor* executor = usage->mutable_executors(i++);
